requirements.txt
fastapi
uvicorn
playwright
pymupdf
anthropic
slack-sdk
requests
python-dotenv
psycopg2-binary
